added implementation + repo + models + unit test for step 2 -simple offers

+++ src/main/scala/com/mini/shop/models/Item.scala
@@ -1,4 +1,4 @@
 package com.mini.shop.models
 
-case class Item(id: Int, name: String, amount: Double)
+case class Item(id: Int, name: String, amount: BigDecimal)
